requirements.txt
Django>=5.0
python-dotenv>=1.0.0
mysqlclient>=2.2.0
djangorestframework>=3.14.0
aiohttp>=3.9.0
requests>=2.31.0
cryptography>=41.0.0
drf-yasg>=1.21.0
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
